Made gun variables loadable from file

+++ src/util.h
@@ -118,7 +118,7 @@ template <typename T> size_t abs_to_index(T value)
 float percentPos(float position, float lowerPos, float higherPos);
 float getPercentPos(float position, float lowerPos, float higherPos);
 
-int stringFindNumber(std::string stringy, std::string term);
+float stringFindNumber(std::string stringy, std::string term);
 std::string stringFindString(std::string stringy, std::string term);
 std::string stringFindStringNpos(std::string stringy, std::string term);
 std::vector<std::string> stringFindElements(std::string source,
